Rename `getDraftCollections` -> `getMyCollections`
